#354 - Fixes in the Web Application.

+++ loa-application/loa-web-application/src/main/java/com/github/bottomlessarchive/loa/web/view/document/service/DocumentStatisticsResponseFactory.java
@@ -97,8 +97,8 @@ public class DocumentStatisticsResponseFactory {
                         .availableStorageInBytes(
                                 serviceInstanceEntity.getProperty("freeSpace")
                                         .map(ServiceInstanceEntityProperty::getValue)
-                                        .map(Integer::parseInt)
-                                        .orElse(-1)
+                                        .map(Long::parseLong)
+                                        .orElse(-1L)
                         )
                         .build()
                 )
